fix: TagSelect Infinite loop
