+++ app/package.json
@@ -20,10 +20,9 @@
     "lodash.get": "^4.4.2",
     "lodash.pick": "^4.4.0",
     "md5": "^2.2.1",
+    "mockjs": "^1.0.2",
     "moment": "^2.24.0",
-    "node-sass": "^4.11.0",
     "nprogress": "^0.2.0",
-    "sass-loader": "^7.1.0",
     "viser-vue": "^2.3.3",
     "vue": "^2.5.22",
     "vue-clipboard2": "^0.2.1",
