+++ app/src/components/TagSelect/TagSelectOption.jsx
@@ -1,111 +1,45 @@
-import PropTypes from 'ant-design-vue/es/_util/vue-types'
-import Option from './TagSelectOption.jsx'
-import { filterEmpty } from '@/components/_util/util'
+import { Tag } from 'ant-design-vue'
+const { CheckableTag } = Tag
 
 export default {
-  Option,
-  name: 'TagSelect',
-  model: {
-    prop: 'checked',
-    event: 'change'
-  },
+  name: 'TagSelectOption',
   props: {
     prefixCls: {
       type: String,
-      default: 'ant-pro-tag-select'
-    },
-    defaultValue: {
-      type: PropTypes.array,
-      default: null
+      default: 'ant-pro-tag-select-option'
     },
     value: {
-      type: PropTypes.array,
-      default: null
+      type: [String, Number, Object],
+      default: ''
     },
-    expandable: {
-      type: Boolean,
-      default: false
-    },
-    hideCheckAll: {
+    checked: {
       type: Boolean,
       default: false
     }
   },
   data () {
     return {
-      expand: false,
-      checkAll: false,
-      localCheckAll: false,
-      items: [],
-      val: this.value || this.defaultValue || []
+      localChecked: this.checked || false
     }
   },
-  methods: {
-    onChange (checked) {
-      const key = Object.keys(this.items).filter(key => key === checked.value)
-      this.items[key] = checked.checked
-      // console.log(this.items)
-      const bool = Object.values(this.items).lastIndexOf(false)
-      console.log('bool', bool, 'this.checkAll', this.checkAll)
-      if (bool === -1) {
-        !this.checkAll && (this.checkAll = true)
-      } else {
-        this.checkAll && (this.checkAll = false)
-        this.localCheckAll = false
-      }
+  watch: {
+    'checked' (val) {
+      this.localChecked = val
     },
-    onCheckAll (checked) {
-      this.checkAll = checked.checked
-      // Object.keys(this.items)
-      // this.items[k] = checked.checked
-      Object.values(this.items).forEach(v => {
-        v = checked.checked
-      })
-    },
-    getItemsKey (items) {
-      const totalItem = {}
-      items.forEach(item => {
-        totalItem[item.componentOptions.propsData && item.componentOptions.propsData.value] = false
-      })
-      return totalItem
-    },
-    // CheckAll Button
-    renderCheckAll () {
-      return !this.hideCheckAll && (<Option key={'total'} checked={this.localCheckAll} onChange={this.onCheckAll}>All</Option>) || null
-    },
-    // expandable
-    renderExpandable () {
-
-    },
-    // render option
-    renderTags (items) {
-      const listeners = {
-        change: (checked) => {
-          this.onChange(checked)
-          this.$emit('change', checked)
-        }
-      }
-
-      return items.map(vnode => {
-        const options = vnode.componentOptions
-        options.listeners = listeners
-        return vnode
-      })
+    '$parent.items': {
+      handler: function (val) {
+        this.value && val.hasOwnProperty(this.value) && (this.localChecked = val[this.value])
+      },
+      deep: true
     }
   },
   render () {
-    const { $props: { prefixCls } } = this
-    const classString = {
-      [`${prefixCls}`]: true
+    const { $slots, value } = this
+    const onChange = (checked) => {
+      this.$emit('change', { value, checked })
     }
-    const tagItems = filterEmpty(this.$slots.default)
-    this.items = this.getItemsKey(tagItems)
-    console.log(this.items)
-    return (
-      <div class={classString}>
-        {this.renderCheckAll()}
-        {this.renderTags(tagItems)}
-      </div>
-    )
+    return (<CheckableTag key={value} vModel={this.localChecked} onChange={onChange}>
+      {$slots.default}
+    </CheckableTag>)
   }
 }
