+++ app/src/components/TagSelect/index.jsx
@@ -1,70 +1,102 @@
 import PropTypes from 'ant-design-vue/es/_util/vue-types'
-import TagSelectOption from './TagSelectOption.jsx'
+import Option from './TagSelectOption.jsx'
 import { filterEmpty } from '@/components/_util/util'
 
 export default {
+  Option,
   name: 'TagSelect',
+  model: {
+    prop: 'checked',
+    event: 'change'
+  },
   props: {
     prefixCls: {
       type: String,
       default: 'ant-pro-tag-select'
     },
     defaultValue: {
-      type: PropTypes.object,
-      default: () => []
-      ,
-      value: {
-        type: PropTypes.object,
-        default: () => []
-        ,
-        expandable: {
-          type: Boolean,
-          default: false
-        },
-        hideCheckAll: {
-          type: Boolean,
-          default: false
-        }
-      },
-      data() {
-        return {
-          expand: false,
-          val: this.value || this.defaultValue || []
-        }
-      },
-      methods: {
-        onChange() {
-
-        },
-
-        // render option
-        renderTag({ value, checked }) {
-          const props = {
-            checked: checked,
-            value: value
-          }
-          return (
-            <TagSelectOption {...{ props }} />
-          )
-        },
-        renderTags(items) {
-          return items.map(item => this.renderTag(item))
-        }
+      type: PropTypes.array,
+      default: null
+    },
+    value: {
+      type: PropTypes.array,
+      default: null
+    },
+    expandable: {
+      type: Boolean,
+      default: false
+    },
+    hideCheckAll: {
+      type: Boolean,
+      default: false
+    }
+  },
+  data () {
+    return {
+      expand: false,
+      localCheckAll: false,
+      items: this.getItemsKey(filterEmpty(this.$slots.default)),
+      val: this.value || this.defaultValue || []
+    }
+  },
+  methods: {
+    onChange (checked) {
+      const key = Object.keys(this.items).filter(key => key === checked.value)
+      this.items[key] = checked.checked
+      const bool = Object.values(this.items).lastIndexOf(false)
+      if (bool === -1) {
+        this.localCheckAll = true
+      } else {
+        this.localCheckAll = false
       }
     },
-    render() {
-      const { $props: { prefixCls }, renderTags } = this
-      const classString = {
-        [`${prefixCls}`]: true
-      }
-      const tagItems = filterEmpty(this.$slots.default)
-      const tagItemDom = (tagItems && tagItems.length > 0) && renderTags(tagItems) || null
+    onCheckAll (checked) {
+      Object.keys(this.items).forEach(v => {
+        this.items[v] = checked.checked
+      })
+    },
+    getItemsKey (items) {
+      const totalItem = {}
+      items.forEach(item => {
+        totalItem[item.componentOptions.propsData && item.componentOptions.propsData.value] = false
+      })
+      return totalItem
+    },
+    // CheckAll Button
+    renderCheckAll () {
+      return !this.hideCheckAll && (<Option key={'total'} checked={this.localCheckAll} onChange={this.onCheckAll}>All</Option>) || null
+    },
+    // expandable
+    renderExpandable () {
 
-      return (
-        <div class={classString}>
-          {tagItemDom}
-        </div>
-      )
+    },
+    // render option
+    renderTags (items) {
+      const listeners = {
+        change: (checked) => {
+          this.onChange(checked)
+          this.$emit('change', checked)
+        }
+      }
+
+      return items.map(vnode => {
+        const options = vnode.componentOptions
+        options.listeners = listeners
+        return vnode
+      })
     }
+  },
+  render () {
+    const { $props: { prefixCls } } = this
+    const classString = {
+      [`${prefixCls}`]: true
+    }
+    const tagItems = filterEmpty(this.$slots.default)
+    return (
+      <div class={classString}>
+        {this.renderCheckAll()}
+        {this.renderTags(tagItems)}
+      </div>
+    )
   }
 }
